+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 lib_sw_pll change log
 =====================
 
+UNRELEASED
+----------
+
+  * ADDED: Fixed frequency settings for 1.536MHz, 3.072MHz and 6.144MHz.
+  * ADDED: Additional sw_pll_result_t return code (SW_PLL_ERR_REG_RW_FAILURE)
+    reported when a switch register read or write fails.
+
 2.5.0
 -----